INFRADEV-11950: Add exception handling for failure in refreshing JIRA issues
